communication with server (user details)

+++ app/src/main/java/hu/bme/aut/android/trackio/data/UserGoals.kt
@@ -1,8 +1,9 @@
 package hu.bme.aut.android.trackio.data
+
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class UserGoals(
-    val goalSteps : Int,
-    val goalWeight : Int
+    val goalSteps: Int,
+    val goalWeight: Float
 )
